requirements.txt
requests
pyyaml
openai
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core